requirements.txt
openai>=1.0.0
streamlit>=1.24.0
pandas>=1.5.0
plotly>=5.13.0
python-dotenv>=0.21.0
pytest>=7.3.1